Rename `parser.py` to `language_parser.py`
This is to avoid confusion with the built-in `parser` module (removed
since Python 3.10).
