+++ src/language/code_generator.cmake
@@ -36,7 +36,7 @@ set(CODE_GENERATOR_PY_FILES
     ${PROJECT_SOURCE_DIR}/src/language/code_generator.py
     ${PROJECT_SOURCE_DIR}/src/language/node_info.py
     ${PROJECT_SOURCE_DIR}/src/language/nodes.py
-    ${PROJECT_SOURCE_DIR}/src/language/parser.py
+    ${PROJECT_SOURCE_DIR}/src/language/language_parser.py
     ${PROJECT_SOURCE_DIR}/src/language/utils.py
 )
